+++ db/migrate/20260909100000_add_reprocessed_to_benefit_overrides.rb
@@ -1,5 +1,0 @@
-class AddReprocessedToBenefitOverrides < ActiveRecord::Migration[8.1]
-  def change
-    add_column :benefit_overrides, :reprocessed, :boolean, default: false, null: false
-  end
-end