+++ spec/factories/appeals.rb
@@ -1,0 +1,7 @@
+FactoryBot.define do
+  factory :appeal do
+    application
+    completed_by factory: [:user]
+    correct { false }
+  end
+end